+++ src/applications/check-in/utils/navigation/pre-check-in/index.js
@@ -4,7 +4,7 @@
  * @param {string} [location.query.id]
  */
 
-import { differenceInHours } from 'date-fns';
+import { updateFormPages } from '..';
 
 const getTokenFromLocation = location => location?.query?.id;
 
@@ -46,13 +46,6 @@ const PRE_CHECK_IN_FORM_PAGES = Object.freeze([
     order: 5,
   },
 ]);
-const now = Date.now();
-
-const isWithInHours = (hours, pageLastUpdated) => {
-  const hoursAgo = differenceInHours(now, pageLastUpdated);
-
-  return hoursAgo <= hours;
-};
 
 const getPagesInOrder = () =>
   [...PRE_CHECK_IN_FORM_PAGES].sort((a, b) => a.order - b.order);
@@ -61,50 +54,15 @@ const createForm = () => {
   return getPagesInOrder().map(page => page.url);
 };
 
-const updateForm = patientDemographicsStatus => {
-  let pages = getPagesInOrder().map(page => page.url);
-  const skippedPages = [];
-  const {
-    demographicsNeedsUpdate,
-    demographicsConfirmedAt,
-    nextOfKinNeedsUpdate,
-    nextOfKinConfirmedAt,
-    emergencyContactNeedsUpdate,
-    emergencyContactConfirmedAt,
-  } = patientDemographicsStatus;
+const updateForm = (patientDemographicsStatus, isEmergencyContactEnabled) => {
+  const pages = PRE_CHECK_IN_FORM_PAGES.map(page => page.url);
 
-  const skipablePages = [
-    {
-      url: URLS.DEMOGRAPHICS,
-      confirmedAt: demographicsConfirmedAt,
-      needsUpdate: demographicsNeedsUpdate,
-    },
-    {
-      url: URLS.NEXT_OF_KIN,
-      confirmedAt: nextOfKinConfirmedAt,
-      needsUpdate: nextOfKinNeedsUpdate,
-    },
-    {
-      url: URLS.EMERGENCY_CONTACT,
-      confirmedAt: emergencyContactConfirmedAt,
-      needsUpdate: emergencyContactNeedsUpdate,
-    },
-  ];
-
-  skipablePages.forEach(page => {
-    const pageLastUpdated = page.confirmedAt
-      ? new Date(page.confirmedAt)
-      : null;
-    if (
-      pageLastUpdated &&
-      isWithInHours(24, pageLastUpdated) &&
-      page.needsUpdate === false
-    ) {
-      skippedPages.push(page.url);
-    }
-  });
-  pages = pages.filter(page => !skippedPages.includes(page));
-  return pages;
+  return updateFormPages(
+    patientDemographicsStatus,
+    isEmergencyContactEnabled,
+    pages,
+    URLS,
+  );
 };
 
 export {
